fix for: issue #27 Wrong Game Size size now stored in bytes and displayed in human readable format. run update_game_sizes.py to update your database if you had v1.2.1 running

+++ modules/models.py
@@ -256,7 +256,7 @@ class Game(db.Model):
     library_name = db.Column(db.String(512), nullable=True)
     library_uuid = db.Column(db.String(36), db.ForeignKey('libraries.uuid'), nullable=False)
 
-    size = db.Column(db.Float, nullable=False, default=0.0)
+    size = db.Column(db.BigInteger, nullable=False, default=0)
 
     def __repr__(self):
         return f"<Game id={self.id}, name={self.name}>"
